move broker_catalog_helper.go functions to the interceptors
